+++ packages/fabric/editor-core/examples/5-full-page.tsx
@@ -106,6 +106,7 @@ export default class Example extends React.Component<Props, State> {
             allowUnsupportedContent={true}
             allowPanel={true}
             allowExtension={true}
+            allowPlaceholderCursor={true}
             mediaProvider={storyMediaProviderFactory()}
             emojiProvider={emojiStoryData.getEmojiResource({
               uploadSupported: true,