+++ packages/fabric/media-filmstrip/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @atlaskit/media-filmstrip
 
+## 5.7.2
+
 ## 5.7.1
 
 ## 5.7.0